Add MF holdings + fix active underlyings LTP + fix fractional ETF qty
Part 1 — Mutual Fund Holdings:
  The 2.19L P&L gap vs Kite combined view was entirely the missing MF
  holding (ABSL Money Manager Fund: 2077.777 units, P&L 2.22L).
  Kite shows a combined equity+MF portfolio; dashboard only had equity.

  adapter: add _get_mf_holdings() via kite.mf_holdings(); get_holdings()
  now returns equity + MF holdings merged.  MF holdings use
  instrument_type="MF" and exchange="MF" to distinguish them downstream.
  Logs the raw MF API row at DEBUG level on first run so field names can
  be verified against the live payload (pledged_quantity field name TBD).

  portfolio_service._inject_ltp_holdings: skip MF holdings — mutual fund
  NAVs are EOD-only; injecting live quotes would find nothing and is
  wasteful.  MF holdings use the last_price (NAV) from the API directly.

  base.get_subscribed_symbols: exclude MF holdings from live feed
  subscription (they have no intraday quotes).

Part 2 — Active Underlyings LTP showing "—":
  RELIANCE options (RELIANCE26MAY1300PE) are open positions; RELIANCE
  equity is NOT in the holdings.  get_subscribed_symbols() only emitted
  the option contract symbol, never the underlying "RELIANCE".
  md_service.get_ltp("RELIANCE") returned None -> LTP chip showed "—".

  Fix: also append p.underlying to subscribed symbols for each position
  when underlying differs from the contract symbol.  This subscribes
  RELIANCE, NIFTY, BANKNIFTY etc. from F&O positions to the quote feed
  so their live prices are available for the underlying chip display.

Part 3 — Fractional ETF quantity truncation:
  Holding.quantity was typed int; adapter cast all quantities with int().
  LIQUIDBEES-F has 0.53 fractional units -> truncated to 0 -> P&L = 0.
  Fixed: Holding.quantity is now float; adapter uses float() throughout.
  This fixes LIQUIDBEES-F and correctly handles any fractional ETF units.

+++ schemas/holding.py
@@ -12,7 +12,7 @@ class Holding:
     symbol: str
     exchange: str           # "NSE" | "BSE"
     isin: str
-    quantity: int
+    quantity: float          # float supports fractional ETF/MF units
     avg_price: float
     ltp: float = 0.0
     current_value: float = 0.0
